getSavedQueriesByStatement

Retrieves a page of the saved queries either created by or shared with the current user. Each SavedQuery in the page, if it is compatible with the current API version, will contain a ReportQuery object which can be optionally modified and used to create a ReportJob. This can then be passed to ReportService runReportJob. The following fields are supported for filtering:

id

name

dfp_getSavedQueriesByStatement(request_data, as_df = TRUE, verbose = FALSE)

Arguments

request_data
a list or data.frame of data elements to be formatted for a SOAP request (XML format, but passed as character string)

as_df
a boolean indicating whether to attempt to parse the result into a data.frame

verbose
a boolean indicating whether to print the service URL and POSTed XML

Value

a data.frame or list containing all the elements of a getSavedQueriesByStatementResponse

See also

Google Documentation for getSavedQueriesByStatement

Examples

# NOT RUN {
request_data <- list(filterStatement=list(query="WHERE id = 936165016"))
this_result <- dfp_getSavedQueriesByStatement(request_data, as_df=FALSE)
this_report_query <- this_result[[1]]$reportQuery

# resubmit the report job with the saved query
report_data <- list(reportJob=list(reportQuery = this_report_query))
report_data <- dfp_full_report_wrapper(report_data)
# }