PublisherQueryLanguageService

Provides methods for executing a PQL Statement to retrieve information from the system. In order to support the selection of columns of interest from various tables, Statement objects support a "select" clause. An example query text might be "select CountryCode, Name from Geo_Target", where CountryCode and Name are columns of interest and Geo_Target is the table.

dfp_select(request_data, verbose = FALSE)

Arguments

request_data
a list or data.frame of data elements to be formatted for a SOAP request (XML format, but passed as character string)

verbose
a boolean indicating whether to print the service URL and POSTed XML

Value

a data.frame or list containing all the elements of a selectResponse

Details

The following tables are supported:

Geo_Target

Bandwidth_Group

Browser

Browser_Language

Device_Capability

Device_Category

Device_Manufacturer

Mobile_Carrier

Mobile_Device

Mobile_Device_Submodel

Operating_System

Operating_System_Version

Third_Party_Company

Line_Item

Ad_Unit

User

Exchange_Rate

Programmatic_Buyer

Audience_Segment_Category

Audience_Segment

Proposal_Retraction_Reason

Audience_Explorer

Time_Zone

Proposal_Terms_And_Conditions

Change_History

Visit the See Also section below to proceed to Google and view columns in each of these tables.

select

Retrieves rows of data that satisfy the given Statement query from the system.

See also

Google Documentation for select

Examples

# NOT RUN {
 request_data <- list(selectStatement=
 list(query='SELECT Id, Name, Targeting FROM LineItem LIMIT 3'))
 dfp_select_result <- dfp_select(request_data)

 request_data <- list(selectStatement=
 list(query="SELECT Id
                  , Name
                  , CanonicalParentId
                  , CountryCode
                  , Type 
             FROM Geo_Target 
             WHERE CountryCode='US' AND (TYPE='STATE' OR TYPE='COUNTY')"))
 us_geos <- dfp_select(request_data)
# }